refactor to content-trust
